// src/app/core/services/model/pointAddressCreateInput.ts
/**
 * TWMS REST API
 * TWMS REST API
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

/**
 * data transfer object for creating a new device point address in frontend representation
 */
export interface PointAddressCreateInput { 
    /**
     * Building of a point address
     */
    building?: string;
    /**
     * Area of a point address
     */
    area?: string;
    /**
     * Room of a point address
     */
    room?: string;
    /**
     * Segment of a point address
     */
    segment?: string;
}